requirements.txt
absl-py
numpy
gemini
python-dotenv

